Move account social request class to account package
